+++ src/main/scala/org/mmisw/orr/ont/app/OrrOntStack.scala
@@ -1,7 +1,5 @@
 package org.mmisw.orr.ont.app
 
-import javax.servlet.http.HttpServletRequest
-
 import com.mongodb.casbah.Imports._
 import org.json4s.JsonAST.{JArray, JNothing, JString, JValue}
 import org.json4s.ext.JodaTimeSerializers
@@ -97,22 +95,6 @@ trait OrrOntStack extends ScalatraServlet with NativeJsonSupport with CorsSuppor
   }
 
   ontUtil.mimeMappings foreach { xm => addMimeMapping(xm._2, xm._1) }
-
-  val defaultRequestedFormat = "json"
-
-  protected def getRequestedFormat(implicit request: HttpServletRequest): String = {
-    def getAcceptHeader = {
-      val ah = acceptHeader
-      //println(s"accept header=$ah")
-      ah
-    }
-    params.get("format") getOrElse (getAcceptHeader match {
-      case List() => defaultRequestedFormat
-      case list =>
-        if (list contains "text/html") "html"
-        else if (mimeTypes.contains(list.head)) mimeTypes(list.head) else defaultRequestedFormat
-    })
-  }
 
   /** Gets a param from the body or from the 'params' */
   protected def getParam(name: String): Option[String] = {